@@ -20,39 +20,39 @@
     "e2e": "ng e2e --webdriver-update=false"
   },
   "dependencies": {
-    "@angular/animations": "~9.1.11",
-    "@angular/cdk": "~9.2.4",
-    "@angular/common": "~9.1.11",
-    "@angular/compiler": "~9.1.11",
-    "@angular/core": "~9.1.11",
-    "@angular/forms": "~9.1.11",
-    "@angular/localize": "~9.1.11",
-    "@angular/material": "~9.2.4",
-    "@angular/platform-browser": "~9.1.11",
-    "@angular/platform-browser-dynamic": "~9.1.11",
-    "@angular/router": "~9.1.11",
-    "@ng-bootstrap/ng-bootstrap": "~6.1.0",
+    "@angular/animations": "~10.0.4",
+    "@angular/cdk": "~10.1.0",
+    "@angular/common": "~10.0.4",
+    "@angular/compiler": "~10.0.4",
+    "@angular/core": "~10.0.4",
+    "@angular/forms": "~10.0.4",
+    "@angular/localize": "~10.0.4",
+    "@angular/material": "~10.1.0",
+    "@angular/platform-browser": "~10.0.4",
+    "@angular/platform-browser-dynamic": "~10.0.4",
+    "@angular/router": "~10.0.4",
+    "@ng-bootstrap/ng-bootstrap": "~7.0.0",
     "bluebird": "~3.7.2",
     "core-js": "~3.6.5",
-    "moment": "~2.26.0",
+    "moment": "~2.27.0",
     "oidc-client": "~1.10.1",
-    "rxjs": "~6.5.5",
-    "rxjs-compat": "~6.5.5",
-    "tslib": "^1.10.0",
+    "rxjs": "~6.6.0",
+    "rxjs-compat": "~6.6.0",
+    "tslib": "^2.0.0",
     "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.8",
-    "@angular/cli": "~9.1.8",
-    "@angular/compiler-cli": "~9.1.11",
-    "@angular/language-service": "~9.1.11",
-    "@types/jasmine": "^3.5.10",
-    "@types/node": "^14.0.13",
-    "codelyzer": "^5.2.2",
+    "@angular-devkit/build-angular": "~0.1000.3",
+    "@angular/cli": "~10.0.3",
+    "@angular/compiler-cli": "~10.0.4",
+    "@angular/language-service": "~10.0.4",
+    "@types/jasmine": "^3.5.11",
+    "@types/node": "^14.0.24",
+    "codelyzer": "^6.0.0",
     "jasmine-core": "^3.5.0",
     "jasmine-reporters": "^2.3.2",
     "jasmine-spec-reporter": "^5.0.2",
-    "karma": "^5.0.9",
+    "karma": "^5.1.0",
     "karma-chrome-launcher": "^3.1.0",
     "karma-cli": "^2.0.0",
     "karma-coverage-istanbul-reporter": "^3.0.3",
@@ -61,6 +61,6 @@
     "protractor": "~7.0.0",
     "ts-node": "^8.10.2",
     "tslint": "^6.1.2",
-    "typescript": "~3.8.3"
+    "typescript": "~3.9.7"
   }
 }
